+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/certifiedProduct/upload/CertifiedProductHandler2015.java
@@ -38,6 +38,8 @@ import gov.healthit.chpl.dto.UcdProcessDTO;
 import gov.healthit.chpl.entity.AddressEntity;
 import gov.healthit.chpl.entity.AttestationType;
 import gov.healthit.chpl.entity.CQMCriterionEntity;
+import gov.healthit.chpl.entity.CertificationCriterionEntity;
+import gov.healthit.chpl.entity.CertificationResultEntity;
 import gov.healthit.chpl.entity.PendingCertificationResultAdditionalSoftwareEntity;
 import gov.healthit.chpl.entity.PendingCertificationResultEntity;
 import gov.healthit.chpl.entity.PendingCertificationResultTestDataEntity;
@@ -1057,10 +1059,19 @@ public class CertifiedProductHandler2015 extends CertifiedProductHandler {
 							cert = certDao.getByName("170.315 (c)(1)");
 						} else if(currCriteria.equals("c2")) {
 							cert = certDao.getByName("170.315 (c)(2)");
+						} else if(currCriteria.equals("c3")) {
+							cert = certDao.getByName("170.315 (c)(3)");
+						} else if(currCriteria.equals("c4")) {
+							cert = certDao.getByName("170.315 (c)(4)");
 						}
 						if(cert != null) {
 							PendingCqmCertificationCriteriaEntity certEntity = new PendingCqmCertificationCriteriaEntity();
 							certEntity.setCertificationId(cert.getId());
+							CertificationCriterionEntity criteriaEntity = new CertificationCriterionEntity();
+							criteriaEntity.setId(cert.getId());
+							criteriaEntity.setNumber(cert.getNumber());
+							criteriaEntity.setTitle(cert.getTitle());
+							certEntity.setCertificationCriteria(criteriaEntity);
 							currResult.getCertifications().add(certEntity);
 						} else {
 							product.getErrorMessages().add("Could not find a certification criteria matching " + currCriteria + " for product " + product.getUniqueId());