+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/manager/impl/PendingCertifiedProductManagerImpl.java
@@ -207,6 +207,8 @@ public class PendingCertifiedProductManagerImpl implements PendingCertifiedProdu
 		//insert the record
 		PendingCertifiedProductDTO pendingCpDto = pcpDao.create(toCreate);
 		updateCertResults(pendingCpDto);
+		//pendingCpDto = pcpDao.findById(pendingCpDto.getId());
+		
 		//add appropriate ACLs
 		//who already has access to this ACB?
 		CertificationBodyDTO acb = acbManager.getById(acbId);
@@ -481,9 +483,9 @@ public class PendingCertifiedProductManagerImpl implements PendingCertifiedProdu
 	public void addAllVersionsToCmsCriterion(PendingCertifiedProductDetails pcpDetails) {
 		//now add allVersions for CMSs
 		String certificationEdition = pcpDetails.getCertificationEdition().get("name").toString();
-		if (certificationEdition.startsWith("2014")){
-			List<CQMCriterion> cqms2014 = getAvailableCQMVersions();
-			for(CQMCriterion cqm : cqms2014) {
+		if (!certificationEdition.startsWith("2011")){
+			List<CQMCriterion> cqms = getAvailableCQMVersions();
+			for(CQMCriterion cqm : cqms) {
 				boolean cqmExists = false;
 				for(CQMResultDetails details : pcpDetails.getCqmResults()) {
 					if(cqm.getCmsId().equals(details.getCmsId())) {
@@ -497,9 +499,10 @@ public class PendingCertifiedProductManagerImpl implements PendingCertifiedProdu
 					result.setNqfNumber(cqm.getNqfNumber());
 					result.setNumber(cqm.getNumber());
 					result.setTitle(cqm.getTitle());
+					result.setDescription(cqm.getDescription());
 					result.setSuccess(Boolean.FALSE);
-					result.setTypeId(cqm.getCqmCriterionTypeId());
 					result.getAllVersions().add(cqm.getCqmVersion());
+					result.setTypeId(cqm.getCqmCriterionTypeId());
 					pcpDetails.getCqmResults().add(result);
 				}
 			}