+++ RELEASE_NOTES.md
@@ -1,5 +1,28 @@
 # Release Notes
 
+## Version 5.0.0
+_19 September 2016_
+
+### Features added
+* Add optional argument 'edition' to /download call.
+  * Generate the chpl-all file as well as a chpl-{edition} file for each edition present in the database for download.
+  * Downloads chpl-all by default
+* Allowed c3/c4 to be connected to CQMs
+* Added error messages for c1/2/3/4-CQM mismatch
+* Re-factored Certification Id web services to improve performance
+* Added statuses object that shows aggregate number of certified products associated with each developer and product. These objects will allow the website search page to filter on a developer/product's number of certified products that are active/retired/withdrawn/suspended/terminated.
+* Updated /developers call to be faster
+* Changed /developers/update to update transparency attestation based on acb name, not acb ID
+* Re-factored Certification Id API controller to support new operations ("create", "search", "verify(GET)")
+* Changed /certification_ids/all operation to just /certification_ids
+
+### Bugs fixed
+* Pending 2015 products can now have CQM versions modified
+* API Documentation Link is now correctly validated on upload and edit 
+* Added new style product number to CAP Activity Report descriptions
+
+---
+
 ## Version 4.0.0
 _30 August 2016_